optimized KeyInput
